added MIT and library doc
