--- flinger/src/main/java/io/iamjosephmj/flinger/flings/DefaultFlingBehavior.kt
@@ -1,32 +1,0 @@
-package io.iamjosephmj.flinger.flings
-
-import androidx.compose.animation.core.*
-import androidx.compose.foundation.gestures.FlingBehavior
-import androidx.compose.foundation.gestures.ScrollScope
-import kotlin.math.abs
-
-class DefaultFlingBehavior(
-    private val flingDecay: DecayAnimationSpec<Float>
-) : FlingBehavior {
-    override suspend fun ScrollScope.performFling(initialVelocity: Float): Float {
-        // come up with the better threshold, but we need it since spline curve gives us NaNs
-        return if (abs(initialVelocity) > 1f) {
-            var velocityLeft = initialVelocity
-            var lastValue = 0f
-            AnimationState(
-                initialValue = 0f,
-                initialVelocity = initialVelocity,
-            ).animateDecay(flingDecay) {
-                val delta = value - lastValue
-                val consumed = scrollBy(delta)
-                lastValue = value
-                velocityLeft = this.velocity
-                // avoid rounding errors and stop if anything is unconsumed
-                if (abs(delta - consumed) > 0.5f) this.cancelAnimation()
-            }
-            velocityLeft
-        } else {
-            initialVelocity
-        }
-    }
-}
