+++ flinger/src/main/java/io/iamjosephmj/flinger/flings/FlingCalculator.kt
@@ -1,7 +1,32 @@
+/*
+* MIT License
+*
+* Copyright (c) 2021 Joseph James
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in all
+* copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*
+*/
+
 package io.iamjosephmj.flinger.flings
 
 import androidx.compose.ui.unit.Density
-import io.iamjosephmj.flinger.ScrollViewConfiguration
+import io.iamjosephmj.flinger.configs.ScrollViewConfiguration
 import io.iamjosephmj.flinger.spline.AndroidFlingSpline
 import kotlin.math.exp
 import kotlin.math.sign
@@ -10,8 +35,10 @@ import kotlin.math.sign
 /**
  * Configuration for Android-feel flinging motion at the given density.
  *
- * @param friction scroll friction.
  * @param density density of the screen. Use LocalDensity to get current density in composition.
+ * @param scrollConfiguration this contain all parameters need for setting the scroll behaviour.
+ *
+ * @author Joseph James
  */
 class FlingCalculator(
     val density: Density,
